+++ src/utils/index.ts
@@ -1,2 +1,5 @@
 export * from "./parser.js";
 export * from "./indexer.js";
+export * from "./logger.js";
+export * from "./cache.js";
+export * from "./downloader.js";